+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentMethodType.swift
@@ -11,7 +11,6 @@ import Foundation
 @_spi(STP) import StripePaymentsUI
 @_spi(STP) import StripeUICore
 import UIKit
-import Stripe
 
 extension PaymentSheet {
     enum PaymentMethodType: Equatable, Hashable {
